app/requirements.txt

# Minimal dependencies required by app/main.py
openai>=1.0.0
streamlit>=1.20.0
python-dotenv>=0.21.0

